# chat/frontend/openapi.json
{
  "components": {
    "schemas": {
      "ChatRequest": {
        "description": "Body for ``POST /api/chat``.\n\nAttributes\n----------\nsession_id\n    Optional. When omitted (or stale) a new session is created with\n    the first 40 characters of ``message`` as its title.\nmessage\n    The user's question. Bounded 1..4000 characters by FastAPI.",
        "properties": {
          "message": {
            "maxLength": 4000,
            "minLength": 1,
            "title": "Message",
            "type": "string"
          },
          "session_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id"
          }
        },
        "required": ["message"],
        "title": "ChatRequest",
        "type": "object"
      },
      "ChatResponse": {
        "description": "Response for ``POST /api/chat``.\n\nMirrors the canonical turn response: answer + citations +\nprovenance fields. ``sql`` is non-null only on the happy path\n(a governed query executed). ``not_answerable=True``\nshort-circuits the SQL + row_count fields.",
        "properties": {
          "answer": {
            "title": "Answer",
            "type": "string"
          },
          "citations": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Citations",
            "type": "array"
          },
          "duration_ms": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Ms"
          },
          "not_answerable": {
            "default": false,
            "title": "Not Answerable",
            "type": "boolean"
          },
          "not_answerable_note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Not Answerable Note"
          },
          "query_ref": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/QueryRef"
              },
              {
                "type": "null"
              }
            ]
          },
          "reasoning_summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reasoning Summary"
          },
          "row_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Row Count"
          },
          "session_id": {
            "title": "Session Id",
            "type": "string"
          },
          "sql": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sql"
          }
        },
        "required": ["session_id", "answer"],
        "title": "ChatResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "QueryRef": {
        "description": "Structured provenance for a governed query's referenced tables.",
        "properties": {
          "source": {
            "enum": ["catalog", "warehouse"],
            "title": "Source",
            "type": "string"
          },
          "tables": {
            "items": {
              "type": "string"
            },
            "title": "Tables",
            "type": "array"
          }
        },
        "required": ["source", "tables"],
        "title": "QueryRef",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": ["loc", "msg", "type"],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Basketball Data Chatbot API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/chat": {
      "post": {
        "description": "Run one chat turn end-to-end and return the final answer.\n\nThin consumer of :func:`chat_server.pipeline.run_turn`: resolves the\nsession id, drains the ``ChatEvent`` generator, and folds the\nevents into one ``ChatResponse``. The cascade (agent \u2192 gate \u2192\ndry-run \u2192 repair \u2192 execute \u2192 compose) lives in exactly one place\n\u2014 ``run_turn`` \u2014 and the JSON and SSE routes are both consumers\nof it.",
        "operationId": "chat_api_chat_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Chat",
        "tags": ["chat"]
      }
    },
    "/api/chat/stream": {
      "post": {
        "description": "SSE endpoint: stream one turn's ``ChatEvent``s back to the UI.\n\nThe endpoint is intentionally thin \u2014 all substantive work happens in\n``chat_server.pipeline.run_turn``. The route is responsible for:\n\n* Resolving the session id (create when the client didn't supply\n  one; mirrors the non-streaming route's behaviour).\n* Mapping each ``ChatEvent`` through ``to_sse_dict`` and emitting\n  one SSE frame.\n* Letting ``EventSourceResponse`` handle the wire-format headers.\n\nErrors after the first event become ``ChatError`` events inside the\nstream \u2014 the response stays 200 (an SSE body that starts with a\n``200 OK`` is well-formed even if a later frame carries an error\ncode). Errors *before* ``turn_started`` would surface as a 500\nbecause the generator never yields; in practice ``run_turn`` yields\n``TurnStarted`` synchronously, so this is a non-issue.",
        "operationId": "chat_stream_api_chat_stream_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Chat Stream",
        "tags": ["chat"]
      }
    },
...
  }
}